refactor: move rv32im callbacks into extensions
