+++ crates/rvr/rvr-openvm-lift/src/cfg.rs
@@ -420,12 +420,7 @@ fn simple_process_instr(instr: &Instr, regs: &mut [Option<u32>; NUM_REGS]) {
             regs[rd as usize] = val;
         }
         // IO / system instructions: most don't write general registers
-        Instr::Store { .. }
-        | Instr::Nop
-        | Instr::HintInput
-        | Instr::PrintStr { .. }
-        | Instr::HintRandom { .. }
-        | Instr::Ext(_) => {}
+        Instr::Store { .. } | Instr::Nop | Instr::Ext(_) => {}
     }
 }
 
@@ -717,12 +712,7 @@ fn transfer_instr(instr: &Instr, state: &mut RegisterState) {
             }
         }
         // IO / system instructions: most don't write general registers
-        Instr::Store { .. }
-        | Instr::Nop
-        | Instr::HintInput
-        | Instr::PrintStr { .. }
-        | Instr::HintRandom { .. }
-        | Instr::Ext(_) => {}
+        Instr::Store { .. } | Instr::Nop | Instr::Ext(_) => {}
     }
 }
