+++ extensions/rv32im/rvr/c/rv32io_callbacks.h
@@ -1,0 +1,13 @@
+#ifndef RV32IO_CALLBACKS_H
+#define RV32IO_CALLBACKS_H
+
+#include <stdint.h>
+
+/* Forwarding stubs owned by the rv32im IO (Rv32Io) extension: the hint-store
+ * consumers (HINT_STOREW, HINT_BUFFER) and REVEAL. Backed by a dispatch table
+ * installed by `Rv32IoExtension::register_host_callbacks`. */
+void openvm_hint_storew(uint32_t dest_addr);
+void openvm_hint_buffer(uint32_t dest_addr, uint32_t num_words);
+void openvm_reveal(uint32_t src_val, uint32_t ptr, uint32_t offset);
+
+#endif /* RV32IO_CALLBACKS_H */
